Create main
This is a todo list I've been working on, incomplete at the current time. Is able to display a list of things to do, organize them by both date and priority. The ability to add subjects to do is available, next I'll be finishing the removal of subjects that the user no longer finds relevant.